+++ eras/allegra/impl/testlib/Test/Cardano/Ledger/Allegra/Examples.hs
@@ -19,6 +19,10 @@ import Cardano.Ledger.Allegra.TxBody
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Genesis (NoGenesis (..))
 import Cardano.Ledger.Shelley.PParams (Update (..))
+import Cardano.Ledger.Shelley.Rules (
+  ShelleyDelegsPredFailure (DelegateeNotRegisteredDELEG),
+  ShelleyLedgerPredFailure (DelegsFailure),
+ )
 import Cardano.Ledger.Shelley.Scripts
 import Cardano.Slotting.Slot
 import Data.Proxy
@@ -44,6 +48,9 @@ import Test.Cardano.Ledger.Shelley.Examples (
 ledgerExamples :: LedgerExamples AllegraEra
 ledgerExamples =
   mkLedgerExamples
+    ( AllegraApplyTxError . pure . DelegsFailure $
+        DelegateeNotRegisteredDELEG @AllegraEra (mkKeyHash 1)
+    )
     (mkWitnessesPreAlonzo (Proxy @AllegraEra))
     exampleCoin
     (exampleAllegraTxBody exampleCoin)