+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/TreeDiff.hs
@@ -40,6 +40,8 @@ import Cardano.Ledger.Dijkstra.Rules (
   DijkstraGovCertPredFailure,
   DijkstraGovPredFailure,
   DijkstraLedgerPredFailure,
+  DijkstraMempoolEvent,
+  DijkstraMempoolPredFailure,
   DijkstraUtxoPredFailure,
   DijkstraUtxowPredFailure,
  )
@@ -197,3 +199,11 @@ instance ToExpr (DijkstraGovCertPredFailure era)
 instance
   ToExpr (PredicateFailure (EraRule "LEDGERS" era)) =>
   ToExpr (DijkstraBbodyPredFailure era)
+
+instance
+  ToExpr (PredicateFailure (EraRule "LEDGER" era)) =>
+  ToExpr (DijkstraMempoolPredFailure era)
+
+instance
+  ToExpr (Event (EraRule "LEDGER" era)) =>
+  ToExpr (DijkstraMempoolEvent era)