+++ libs/ledger-state/bench/Performance.hs
@@ -11,6 +11,7 @@ import Cardano.Ledger.Address
 import Cardano.Ledger.Api.Era
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Binary
+import Cardano.Ledger.Conway
 import Cardano.Ledger.Conway.Rules (
   ConwayLedgerPredFailure (ConwayUtxowFailure),
   ConwayUtxowPredFailure (InvalidWitnessesUTXOW),
@@ -79,7 +80,7 @@ main = do
       !globals = mkGlobals genesis
       !slotNo = SlotNo 55733343
       restrictError = \case
-        ApplyTxError (ConwayUtxowFailure (InvalidWitnessesUTXOW _) :| []) -> ()
+        ConwayApplyTxError (ConwayUtxowFailure (InvalidWitnessesUTXOW _) :| []) -> ()
         otherErr -> error . show $ otherErr
       applyTx' mempoolEnv mempoolState =
         -- TODO: revert this to `either (error . show) seqTuple` after tx's are fixed